encapsulate gem inside protobuf/activerecord folder and require protobuf explicitly
