+++ lib/protobuf-activerecord.rb
@@ -1,5 +1,5 @@
-require 'protobuf_ext/message'
-
-require 'protoable'
-
+require 'heredity'
+require 'protobuf'
 require 'protobuf/activerecord/version'
+require 'protobuf/activerecord/protobuf_ext/message'
+require 'protobuf/activerecord/protoable'
