+++ lib/protobuf/activerecord/protoable/convert.rb
