+++ lib/protobuf/activerecord/protoable/errors.rb
