+++ lib/protobuf/activerecord/protoable/fields.rb
