+++ lib/protobuf/activerecord/protoable/persistence.rb
