+++ lib/protobuf/activerecord/protoable/scope.rb
