+++ lib/protobuf/activerecord/protoable/serialization.rb
